+++ CHANGELOG.md
@@ -1,5 +1,19 @@
-# 1.2.10
+# next
+
+- Update `onSkip` in `README`
+- Fix issue [192](https://github.com/RafaelBarbosatec/tutorial_coach_mark/issues/192)
+
+# 1.2.12
 - Change moment where call `_controller.forward()` occurs. That prevents the `unFocusAnimationDuration` is always called;
+
+# 1.2.11
+- Fix issue [165](https://github.com/RafaelBarbosatec/tutorial_coach_mark/issues/165)
+
+# 1.2.10
+- now `onSkip` return a bool. if onSkip return false, the overlay will not be dismissed and call `next`. [#89](https://github.com/RafaelBarbosatec/tutorial_coach_mark/issues/89)
+- Able to set to use or not SafeArea widget. Thanks [devmasteripatas](https://github.com/devmasteripatas).
+- Adds `initialFocus` in `TutorialCoachMark` to select witch focus you want init.
+- Adds method `goTo(index)` in `TutorialCoachMark` to navigate to focus you want.
 
 # 1.2.9
 - set `ignoringSemantics` false in `skip` button. Reolve [#156](https://github.com/RafaelBarbosatec/tutorial_coach_mark/issues/156)